WIP New script: "Hardware"
0 dependancy GPU, CPU, RAM monitor

GPU likely only works with AMDGPU

Cpu % is a bit strange looking but I mean it seems to work.

RAM Seems fine but is higher than reported by other programs.
Only reads MemAvailable which is afaik used for new allocations so that
should be the important part

Seems decent. Not really ready for a full release yet, lots of
hardcoded/temporary values